backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-core==2.20.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
cachetools==5.3.3
loguru==0.7.2
# Google APIs (calendar, later)
google-api-python-client==2.137.0
google-auth==2.32.0
langchain_openai
langchain_community
langchain_mcp_adapters
langchain-core
langgraph-supervisor
langgraph==0.5.3
langgraph-prebuilt==0.5.2
